refactor: authorization 스크립트 모듈화
