--- js/auth/signin.js
@@ -1,4 +1,0 @@
-import { initSignin } from "./functions.js";
-
-const init = initSignin;
-init();
