--- js/auth/signup.js
@@ -1,4 +1,0 @@
-import { initSignup } from "./functions.js";
-
-const init = initSignup;
-init();
