+++ js/signin.js
@@ -1,0 +1,4 @@
+import { initSignin } from "./authorization.js";
+
+const init = initSignin;
+init();
