+++ js/signup.js
@@ -1,0 +1,4 @@
+import { initSignup } from "./authorization.js";
+
+const init = initSignup;
+init();
