+++ pages/signup.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Linkbrary</title>
     <link rel="stylesheet" href="/css/main.css" />
-    <script type="module" src="/js/auth/signup.js" defer></script>
+    <script type="module" src="/js/signup.js" defer></script>
   </head>
   <body>
     <div class="container">
